requirements.txt
fastapi
uvicorn
requests
python-dotenv
scikit-learn
pandas
joblib